requirements.txt
torch
torchvision
torchaudio
--index-url https://download.pytorch.org/whl/cu124numpy
opencv-python
pillow
albumentations
tqdm